plotter: move some int drawing to floating point

+++ src/qtgui/plotter.h
@@ -322,7 +322,7 @@ private:
 
     // Waterfall averaging
     quint64     tlast_wf_ms;        // last time waterfall has been updated
-    quint64     msec_per_wfline;    // milliseconds between waterfall updates
+    double     msec_per_wfline;    // milliseconds between waterfall updates
     quint64     wf_span;            // waterfall span in milliseconds (0 = auto)
     int         fft_rate;           // expected FFT rate (needed when WF span is auto)
 };
